# bot/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dYdX Trading Bot API",
    "description": "API for managing multiple dYdX trading bot instances with JWT Authentication. Most non-auth HTTP endpoints return a standardized envelope: {success, message, data, timestamp, trace_id}.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/bots/{instance_id}": {
      "get": {
        "summary": "Get Bot Instance",
        "description": "Get specific bot instance status",
        "operationId": "get_bot_instance_api_v1_bots__instance_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BotInstanceStatus"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      },
      "delete": {
        "summary": "Delete Bot Instance",
        "description": "Delete bot instance",
        "operationId": "delete_bot_instance_api_v1_bots__instance_id__delete",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
    "/api/v1/bots/{instance_id}/history": {
      "get": {
        "summary": "Get Bot History",
        "description": "Get bot event history",
        "operationId": "get_bot_history_api_v1_bots__instance_id__history_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
    "/api/v1/bots/{instance_id}/jobs": {
      "get": {
        "summary": "Get Bot Jobs",
        "description": "Get bot job history",
        "operationId": "get_bot_jobs_api_v1_bots__instance_id__jobs_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
    "/api/v1/bots/{instance_id}/trades": {
      "get": {
        "summary": "Get Bot Trades",
        "description": "Get bot trades, newest first.\n\n``limit`` and ``offset`` page through the trades left after the ``status``\nfilter; without ``limit`` (or with ``limit`` <= 0) every matching trade is\nreturned. ``total_trades`` counts the matching trades before paging and\n``count`` the trades in this page.",
        "operationId": "get_bot_trades_api_v1_bots__instance_id__trades_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
    "/api/v1/bots/{instance_id}/cointegrated-pairs": {
      "get": {
        "summary": "Get Bot Cointegrated Pairs",
        "description": "Get the cointegrated pairs the bot's last pair scan stored.\n\nReads the ``cointegrated_pairs`` row the worker wrote for this instance:\n``analyzed_at`` is when that scan ran and ``pairs`` carries its statistics\nper pair. Both are empty when the instance never ran with pair discovery.\nStatistics only; no account data.",
        "operationId": "get_bot_cointegrated_pairs_api_v1_bots__instance_id__cointegrated_pairs_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
    "/api/v1/bots/{instance_id}/stats": {
      "get": {
        "summary": "Get Bot Stats",
        "description": "Get bot statistics",
        "operationId": "get_bot_stats_api_v1_bots__instance_id__stats_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {}
                      }
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-response-envelope": "StandardApiResponse"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StandardApiResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data",
          "timestamp",
          "trace_id"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": [
              "object",
              "array",
              "string",
              "number",
              "boolean",
              "null"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "trace_id": {
            "type": "string"
          }
        },
        "description": "Standard API response envelope used by bot/runtime HTTP endpoints."
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    }
  }
}